Fixed some PC-related tests. Did some documentation for the external algorithm API.
