+++ tetrad-lib/src/test/java/edu/cmu/tetrad/test/joes/examples/ExampleCompareFromFilesSimple.java
@@ -19,14 +19,11 @@
 // Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA //
 ///////////////////////////////////////////////////////////////////////////////
 
-package edu.cmu.tetrad.algcomparison.joe_examples.examples;
+package edu.cmu.tetrad.test.joes.examples;
 
 import edu.cmu.tetrad.algcomparison.Comparison;
 import edu.cmu.tetrad.algcomparison.algorithm.Algorithms;
-import edu.cmu.tetrad.algcomparison.algorithm.oracle.pattern.*;
-import edu.cmu.tetrad.algcomparison.independence.FisherZ;
-import edu.cmu.tetrad.algcomparison.algorithm.ExternalAlgorithm;
-import edu.cmu.tetrad.algcomparison.score.FisherZScore;
+import edu.cmu.tetrad.algcomparison.algorithm.external.ExternalAlgorithmPcalgPc;
 import edu.cmu.tetrad.algcomparison.statistic.*;
 import edu.cmu.tetrad.util.Parameters;
 
@@ -41,17 +38,14 @@ import edu.cmu.tetrad.util.Parameters;
  *
  * @author jdramsey
  */
-public class ExampleCompareFromFiles {
+public class ExampleCompareFromFilesSimple {
     public static void main(String... args) {
         Parameters parameters = new Parameters();
 
         // Can leave the simulation parameters out since
         // we're loading from file here.
-        parameters.set("alpha", 0.001, 1e-8);
+        parameters.set("alpha", 1e-3);
         parameters.set("numRuns", 10);
-        parameters.set("penaltyDiscount", 4);
-        parameters.set("useMaxPOrientationHeuristic", true);
-        parameters.set("maxPOrientationMaxPathLength", 3);
 
         Statistics statistics = new Statistics();
 
@@ -61,13 +55,12 @@ public class ExampleCompareFromFiles {
         statistics.add(new AdjacencyRecall());
         statistics.add(new ArrowheadPrecision());
         statistics.add(new ArrowheadRecall());
-        statistics.add(new PercentBidirectedEdges());
-//        statistics.add(new MathewsCorrAdj());
-//        statistics.add(new MathewsCorrArrow());
-//        statistics.add(new F1Adj());
-//        statistics.add(new F1Arrow());
-//        statistics.add(new SHD());
-//        statistics.add(new ElapsedTime());
+        statistics.add(new MathewsCorrAdj());
+        statistics.add(new MathewsCorrArrow());
+        statistics.add(new F1Adj());
+        statistics.add(new F1Arrow());
+        statistics.add(new SHD());
+        statistics.add(new ElapsedTime());
 
         statistics.setWeight("AP", 1.0);
         statistics.setWeight("AR", 0.5);
@@ -76,23 +69,17 @@ public class ExampleCompareFromFiles {
 
         Algorithms algorithms = new Algorithms();
 
-        algorithms.add(new Pc(new FisherZ()));
-        algorithms.add(new PcStableMax(new FisherZ(), false));
-        algorithms.add(new Cpc(new FisherZ()));
-        algorithms.add(new CpcStable(new FisherZ()));
-        algorithms.add(new PcStable(new FisherZ()));
-        algorithms.add(new CpcStable(new FisherZ()));
-        algorithms.add(new Fges(new FisherZScore()));
-
-        algorithms.add(new ExternalAlgorithm("rgraph"));
+//        algorithms.add(new Pc(new FisherZ()));
+        algorithms.add(new ExternalAlgorithmPcalgPc("rgraph"));
 
         Comparison comparison = new Comparison();
         comparison.setShowAlgorithmIndices(true);
         comparison.setShowSimulationIndices(true);
-        comparison.setSortByUtility(false);
-        comparison.setShowUtilities(false);
+        comparison.setSortByUtility(true);
+        comparison.setShowUtilities(true);
+        comparison.setParallelized(true);
 
-        comparison.compareFromFiles("comparison", algorithms, statistics, parameters);
+        comparison.compareFromFiles("comparison_simple", algorithms, statistics, parameters);
     }
 }
