+++ tetrad-lib/src/test/java/edu/cmu/tetrad/test/joes/examples/ExampleCompareSimulation.java
@@ -19,38 +19,36 @@
 // Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA //
 ///////////////////////////////////////////////////////////////////////////////
 
-package edu.cmu.tetrad.algcomparison.joe_examples.examples;
+package edu.cmu.tetrad.test.joes.examples;
 
 import edu.cmu.tetrad.algcomparison.Comparison;
 import edu.cmu.tetrad.algcomparison.algorithm.Algorithms;
-import edu.cmu.tetrad.algcomparison.algorithm.ExternalAlgorithm;
+import edu.cmu.tetrad.algcomparison.algorithm.oracle.pattern.*;
+import edu.cmu.tetrad.algcomparison.graph.RandomForward;
+import edu.cmu.tetrad.algcomparison.independence.FisherZ;
+import edu.cmu.tetrad.algcomparison.score.SemBicScore;
+import edu.cmu.tetrad.algcomparison.simulation.SemSimulation;
+import edu.cmu.tetrad.algcomparison.simulation.Simulations;
 import edu.cmu.tetrad.algcomparison.statistic.*;
 import edu.cmu.tetrad.util.Parameters;
 
 /**
- * An example script to load in data sets and graphs from files and analyze them. The
- * files loaded must be in the same format as
- * </p>
- * new Comparison().saveDataSetAndGraphs("comparison/save1", simulation, parameters);
- * </p>
- * saves them. For other formats, specialty data loaders can be written to implement the
- * Simulation interface.
+ * An example script to simulate data and run a comparison analysis on it.
  *
  * @author jdramsey
  */
-public class ExampleCompareFromFilesSimple {
+public class ExampleCompareSimulation {
     public static void main(String... args) {
         Parameters parameters = new Parameters();
 
-        // Can leave the simulation parameters out since
-        // we're loading from file here.
-        parameters.set("alpha", 1e-3);
         parameters.set("numRuns", 10);
+        parameters.set("numMeasures", 100);
+        parameters.set("avgDegree", 4, 6);
+        parameters.set("sampleSize", 500);
+        parameters.set("alpha", 1e-4, 1e-3, 1e-2);
 
         Statistics statistics = new Statistics();
 
-        statistics.add(new ParameterColumn("avgDegree"));
-        statistics.add(new ParameterColumn("sampleSize"));
         statistics.add(new AdjacencyPrecision());
         statistics.add(new AdjacencyRecall());
         statistics.add(new ArrowheadPrecision());
@@ -64,22 +62,27 @@ public class ExampleCompareFromFilesSimple {
 
         statistics.setWeight("AP", 1.0);
         statistics.setWeight("AR", 0.5);
-        statistics.setWeight("AHP", 1.0);
-        statistics.setWeight("AHR", 0.5);
 
         Algorithms algorithms = new Algorithms();
 
-//        algorithms.add(new Pc(new FisherZ()));
-        algorithms.add(new ExternalAlgorithm("rgraph"));
+        algorithms.add(new Pc(new FisherZ()));
+        algorithms.add(new Cpc(new FisherZ(), new Fges(new SemBicScore(), false)));
+        algorithms.add(new PcStable(new FisherZ()));
+        algorithms.add(new CpcStable(new FisherZ()));
+
+        Simulations simulations = new Simulations();
+
+        simulations.add(new SemSimulation(new RandomForward()));
 
         Comparison comparison = new Comparison();
+
         comparison.setShowAlgorithmIndices(true);
         comparison.setShowSimulationIndices(true);
         comparison.setSortByUtility(true);
         comparison.setShowUtilities(true);
         comparison.setParallelized(true);
 
-        comparison.compareFromFiles("comparison_simple", algorithms, statistics, parameters);
+        comparison.compareFromSimulations("comparison", simulations, algorithms, statistics, parameters);
     }
 }
