+++ tetrad-lib/src/test/java/edu/cmu/tetrad/test/joes/examples/RunConfig.java
@@ -19,52 +19,18 @@
 // Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA //
 ///////////////////////////////////////////////////////////////////////////////
 
-package edu.cmu.tetrad.algcomparison.joe_examples.examples;
+package edu.cmu.tetrad.test.joes.examples;
 
 import edu.cmu.tetrad.algcomparison.Comparison;
-import edu.cmu.tetrad.algcomparison.graph.RandomForward;
-import edu.cmu.tetrad.algcomparison.simulation.LinearSineSimulation;
-import edu.cmu.tetrad.algcomparison.simulation.Simulation;
-import edu.cmu.tetrad.util.Parameters;
 
 /**
  * An example script to save out data files and graphs from a simulation.
  *
  * @author jdramsey
  */
-public class ExampleNonlinearSave {
+public class RunConfig {
     public static void main(String... args) {
-        Parameters parameters = new Parameters();
-
-        parameters.set("numRuns", 100);
-        parameters.set("numMeasures", 20);
-        parameters.set("avgDegree", 3);
-        parameters.set("sampleSize", 500);
-        parameters.set("percentDiscrete", 0);
-        parameters.set("minCategories", 2);
-        parameters.set("maxCategories", 5);
-        parameters.set("differentGraphs",true);
-
-        parameters.set("interceptLow", 0);
-        parameters.set("interceptHigh", 1);
-        parameters.set("contiuousInfluence", 0.5);
-        parameters.set("linearLow", 0.5);
-        parameters.set("linearHigh", 1.0);
-        parameters.set("quadraticLow", 0.5);
-        parameters.set("quadraticHigh", 1.0);
-        parameters.set("cubicLow", 0.2);
-        parameters.set("cubicHigh", 0.3);
-        parameters.set("varLow", 1);
-        parameters.set("varHigh", 1);
-        parameters.set("betaLow", 5);
-        parameters.set("betaHigh", 8);
-        parameters.set("gammaLow", 1.0);
-        parameters.set("gammaHigh", 1.5);
-
-        Simulation simulation = new LinearSineSimulation(new RandomForward());
-        Comparison comparison = new Comparison();
-        comparison.setShowAlgorithmIndices(true);
-        comparison.saveToFiles("comparison", simulation, parameters);
+        new Comparison().configuration("comparison/Config.txt");
     }
 }
